requirements.txt
Django==4.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cloudinary-storage==0.3.0
cloudinary==1.44.1
Pillow==12.0.0
requests==2.32.5